Centralize the snapshot tests for form under core To minimize the duplication of snapshot testing code across themes, centralized the array ones in core in a manner similar to how schema tests work with utils and the validators. Also added snapshot tests in core for object (forgotten in last PR) and form
