+++ packages/core/package.json
@@ -11,6 +11,7 @@
     "publish-to-npm": "npm run build && npm publish",
     "test": "dts test",
     "test:debug": "node --inspect-brk node_modules/.bin/dts test",
+    "test:update": "dts test --u",
     "test:watch": "dts test --watch",
     "test-coverage": "dts test --coverage"
   },
