+++ packages/core/test/FormSnap.test.jsx
@@ -1,0 +1,4 @@
+import Form from '../src';
+import formTests from '../testSnap/formTests';
+
+formTests(Form);
