+++ packages/core/test/ObjectSnap.test.jsx
@@ -1,0 +1,4 @@
+import Form from '../src';
+import objectTests from '../testSnap/objectTests';
+
+objectTests(Form);
